Group various 'ancien_tiers_etat' decls and defs together
Readelf confirms that both variables are completely separate: modifying
'ancien_tiers_etat' in etats_calcul.c is never seen by code in
etats_affiche.c and vice versa.

This is probably not what was intended.

For now, I'll leave this variable alone until we figure out what to do with
it. I'm just moving the decls around to make the issue more obvious when
we revisit the issue later on.