tests: fix gsb_real/gsb_data_account related tests
All the literal integers now wrapped in the G_GINT64_CONSTANT were being
compiled as another value as they overflow traditional 32bit integer
limits.

I'm not quite sure why this is failing now, but this is the correct way to
specify 64bit literal integers as the LL prefix is a GNU extension.