Fix 2 leaks of GSList "sup_payees" in tiers_onglet.c
This list is created and updated when running the payee editing/merging
wizard. It holds the id numbers of payees that will ultimately be
removed. This GSList is stored as a data property of the gsb_assitant
widget initialized in payees_manage_payees().

This patch makes sure that the list is freed before being overwritten or
when its parent object is freed.
(cherry picked from commit 3b7e6062a32df24fa0a063b6c5b6f05bbe66311c)