Fix huge memory leak in gtk_combofix.c
GtkWindows are referenced by Gtk's top-level widget list.

GtkTreeStore, GtkTreeModelFilter and GtkTreeModelSort are all derived
from GObject (and not GInitiallyUnowned like all widgets) so they are
created with a reference of one. We need to explicitly unref them.

The store and model filters hold large amounts of memory: proportional to
the number of categories and payees. In my case, this represents around
300 KB.