Fix leak of 'navigation_model' gsb_gui_navigation_create_navigation_pane()
GObject-derived types are created with a reference count of 1 (unlike
GInitiallyUnowned-derived types which are created with a floating ref)
so they need to be unref()ed explicitly at some point.

In this case, the fix is correct but not satisfactory as we keep using
the 'global' static pointer without properly holding a reference.