Fix a use-after-free() crash when loading a second file
This bug was brought to light following 8dc5a23

In the function gsb_file_load_logo_accueil(),
gsb_select_icon_create_pixbuf_from_chaine_base64() creates a brand new
pixbuf with a reference count of 1. gtk_window_set_default_icon() then
sets it to 2.

The bug is in gsb_select_icon_set_logo_pixbuf() which sometimes creates
a new (resized) pixbuf from the first one, or sometimes just keeps a
pointer to the previous one.

When loading the second file, gtk_window_set_default() is called again
with a new pixbuf and it unrefs the old pixbuf. At this point, the
reference count drops to 0 and the pixbuf is freed.
gsb_select_icon_set_logo_pixbuf() is called again and the old pixbuf is
unref()ed, thus causing an access to freed memory.