i18n: add src/gsb_rgba.c in the list of l10n files
As suggested by "make check":
The following files contain translations and are currently not in use.
Please consider adding these to the POTFILES.in file, located in the po/
directory.

src/gsb_rgba.c

If some of these files are left out on purpose then please add them to
POTFILES.skip instead of POTFILES.in. A file 'missing' containing this
list of left out files has been written in the current directory.