AppVeyor integration, 32 & 64bit
* MSYSTEM used in AppVeyor's environment (ease integration)
* Update of MSYS2 core system disabled (may block AppVeyor as it may require to start a new console)
* First import for install/builder script
* Support of Windows 32bit & 64bit targets
* Explicit version & architecture in artifacts

In appveyor.yml:
Delivery keys have to be adjusted; also, note that only autotools version (currently 1.1.1) is kept in AppVeyor: no build number is used for now, so that AppVeyor may complain about duplicate version. As a workaround, remove any duplicate version(s) in AppVeyor and in github's packages' section.