Upstream changes from 21 to 27 march (#2)
* fixed bug 1829: incorrect display of the full file name
 displays the basename of the file in the button and the long name in a tooltip
* macOS: add missing icons for +/- in lists
These icons are often used in the Preferences window.
* fixes display bug of the last scheduled transaction of the period
fixes display bug of the last scheduled transaction when its date is the last day of the period
* Added all scheduled transactions of the month if conf.execute_scheduled_of_month = TRUE
* New version of the grisbi file.
Adding Fd (fixed day) data in scheduler transaction
* release unstable version 1.1.92
* fixes a warning when displaying reconciled transactions
* remove temporary code to fix crash during the transition to gtk-2.24-30.
* Modified the Cancel action in the "Quit Grisbi" window
The "Cancel" button only cancels the closure of Grisbi.
* revert the commit 906ad17
Operation is impractical and more inconvenient than the previous solution
* Updates german strings