Fix use of sizeof()
According to g_utf8_strncpy() documentation:
https://developer.gnome.org/glib/stable/glib-Unicode-Manipulation.html#g-utf8-strncpy

Note you must ensure dest is at least 4 * n to fit the largest possible
UTF-8 characters

I don't know why sizeof (char*) was used here.
Luckily sizeof(char*) is 4 on a 32-bits system and 8 on a 64-bits system
so this bug was not the source of crash.