Fix a crash in new account assistant
A crash occured when the "Name and accouting entity" was edited.

The crash occurred in g_object_set_data() because the previous data was
already free-ed. The deallocation is done automatically by
g_object_set_data() because the data has been stored using
g_object_set_data_full() with an explicit destructor.