autogen.sh: check for libtoolize version
Since the script uses libtoolize we check its version and not the
version of libtool.
libtool is not used for building Grisbi.

libtoolize may be installed and libtool not installed. The autogen.sh
script would then fails for nothing.

On Debian libtoolize is in the libtool package and libtool is in the
libtool-bin package.