Disable QUIT while saving a file
The problem is that if you are fast enough to do Ctrl-S and Ctrl-Q (and
the save is slow) the Quit action was executed before the Save finished.
The Save function modifies some GUI status but the window was already
closed because of Quit. And we get a crash.

Now the Quit action is ignored while a Save or Save-as is executing.

Fixes bug #1969
http://www.grisbi.org/bugsreports/view.php?id=1969