Cancel the exit if Esc key is used in the save diablog
If the file is modified and "save on exit" is not enabled then we have a
dialog asking "Save changes to document '%s' before closing?" with 3
buttons: "Close without saving", "cancel", "save"

There is also another option, use the Esc key to abort the dialog.
This case was handled as "save". It is now handled as "cancel" and
Grisbi is not exited.