Escape the characters of a file name before display
If the filename contains a "&" and savig the file fails (for example the
directory to save the file does not exist) then the gtk+3 dialog
contained nothing expect the Close button.

The special characters must be escaped before they are sent as a Pango
Markup Language input in gtk_message_dialog_set_markup() used by
dialogue_special().

Closes https://www.grisbi.org/bugsreports/view.php?id=2089