macOS: fix start of Grisbi by double click on a .gsb file
If Grisbi is configured to open the last opened file and is started by
double click on a .gsb file then Grisbi opened 2 files at the same time.
The display was split in 2 with 2 accompts displayed at the same time.

Now a double click to start Grisbi is seen as if Grisbi is started with
the filename as argument, like on GNU/Linux.