Fix a crash in utils_str_break_form_name_field()
Do not use a special treatment for the empty string.
The problem was that index i was outside of the bound of tab[].

Problematic string: "Dépenses vacances Marseille avril-mai 2016 "
with trunc=20.r
Note the space character at the end of the string.

Thanks to Jean-Luc Duflot for the bug report.