Display the message "This is not a Grisbi file" only once
We should not display the error "This is not a Grisbi file" for every
token found in the file. Or the user wil have to click to close the
error dialog an insane amount of times.

An easy way to reproduce the problem is to edit a .gsb file and replace
the tag <Grisbi> by anything else like <foo>.