Use devel_debug() instead of devel_print_str()
devel_print_str() is used to send the log string in the log file only.
The debug message was NOT sent to the console even when --debug= is used.

devel_debug() will not display the message on the console if --debug= is
used and also in the log file if the debug mode is enabled.