Do not display 2 dialogs asking for orphan childer removal
If some scheduled paiments are orphan we get the dialog asking what to do
twice. We get the second dialog while the first one is still displayed.
This is because the same code is called from 2 Gtk+ origins and the
dialog is not really modal.

Now the dialog is displayed only once at a time.