Fix a crash in CVS import
The global variable csv_create_rule was never reset to FALSE.
So once set to TRUE any further CVS import was supposed to have rules
associated. But if no rule was created then the import crashed.

Now we only use the presence (or not) of the "csv-import-rule" data to
know if a rule must be created or not.